LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/worker-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: Worker

FAIL Base URL in workers: new Worker() assert_unreached: Got error event Reached unreachable code
